+++ humility-core/src/core/mod.rs
@@ -29,8 +29,8 @@ pub trait Core {
     fn info(&self) -> (String, Option<String>);
     fn read_word_32(&mut self, addr: u32) -> Result<u32>;
     fn read_8(&mut self, addr: u32, data: &mut [u8]) -> Result<()>;
-    fn read_reg(&mut self, reg: Register) -> Result<u32>;
-    fn write_reg(&mut self, reg: Register, value: u32) -> Result<()>;
+    fn read_reg(&mut self, reg: Register) -> Result<u64>;
+    fn write_reg(&mut self, reg: Register, value: u64) -> Result<()>;
     fn init_swv(&mut self) -> Result<()>;
     fn read_swv(&mut self) -> Result<Vec<u8>>;
     fn write_word_32(&mut self, addr: u32, data: u32) -> Result<()>;
